improvements to post processing pipeline, view update runs on a thread, FXAA and begin on SMAA

+++ include/zg/components/windows/FXAA.hpp
@@ -2,5 +2,5 @@
 #include "WindowComponent.hpp"
 namespace zg::components::windows
 {
-    WindowComponentCreateInfo FXAAFactory(float edgeThresholdMin = 0.0312f, float edgeThreshold = 0.125f, int edgeSearchSteps = 12, float subpixQuality = 0.75f);
+    WindowComponentCreateInfo FXAAFactory(float edgeThresholdMin = 0.0312f, float edgeThreshold = 0.125f, float edgeSearchSteps = 12, float subpixQuality = 0.75f);
 }
